# docs/application-openapi.yaml
openapi: 3.1.0
info: {title: Tailors SaaS Application API, version: 1.0.0, description: Tenant-scoped web application API. The dedicated voice contract remains in openapi.yaml.}
servers: [{url: https://api.tailors.app/api/v1}]
security: [{sanctumSession: []}]
paths:
...
  /purchases: {get: {operationId: listPurchases, parameters: [{$ref: '#/components/parameters/Cursor'}], responses: {'200': {$ref: '#/components/responses/Success'}}}, post: {operationId: receivePurchase, requestBody: {$ref: '#/components/requestBodies/Purchase'}, responses: {'201': {$ref: '#/components/responses/Success'}}}}
...
  /sales: {get: {operationId: listSales, parameters: [{$ref: '#/components/parameters/Cursor'}], responses: {'200': {$ref: '#/components/responses/Success'}}}, post: {operationId: completeSale, requestBody: {$ref: '#/components/requestBodies/Sale'}, responses: {'201': {$ref: '#/components/responses/Success'}}}}
...
  /expenses/{expense}/attachments: {post: {operationId: uploadExpenseAttachment, parameters: [{$ref: '#/components/parameters/ExpenseId'}], responses: {'201': {$ref: '#/components/responses/Success'}}}}
  /expenses/{expense}/approve: {post: {operationId: approveExpense, parameters: [{$ref: '#/components/parameters/ExpenseId'}], responses: {'200': {$ref: '#/components/responses/Success'}}}}
  /expense-attachments/{attachment}: {get: {operationId: downloadExpenseAttachment, parameters: [{$ref: '#/components/parameters/AttachmentId'}], responses: {'200': {description: Private attachment, content: {application/octet-stream: {schema: {type: string, format: binary}}}}}}}
...
  /rentals: {get: {operationId: listRentals, parameters: [{$ref: '#/components/parameters/Cursor'}], responses: {'200': {$ref: '#/components/responses/Success'}}}, post: {operationId: reserveRental, requestBody: {$ref: '#/components/requestBodies/Rental'}, responses: {'201': {$ref: '#/components/responses/Success'}}}}
  /rentals/{rental}/return: {post: {operationId: returnRental, parameters: [{$ref: '#/components/parameters/RentalId'}], requestBody: {$ref: '#/components/requestBodies/RentalReturn'}, responses: {'200': {$ref: '#/components/responses/Success'}}}}
...
  /exports/{export}/download: {get: {operationId: downloadExport, parameters: [{$ref: '#/components/parameters/ExportId'}], responses: {'200': {description: Generated export}}}}
...
  /tenant-deletion/{deletion}: {delete: {operationId: cancelDeletion, parameters: [{$ref: '#/components/parameters/DeletionId'}], responses: {'200': {$ref: '#/components/responses/Success'}}}}
...
components:
  securitySchemes:
    sanctumSession: {type: apiKey, in: cookie, name: laravel_session, description: Obtain with /sanctum/csrf-cookie; mutations require X-XSRF-TOKEN.}
  parameters:
    Cursor: {name: cursor, in: query, required: false, schema: {type: string}}
...
    ExpenseId: {name: expense, in: path, required: true, schema: {type: string, minLength: 26, maxLength: 26}}
    AttachmentId: {name: attachment, in: path, required: true, schema: {type: string, minLength: 26, maxLength: 26}}
    RentalId: {name: rental, in: path, required: true, schema: {type: string, minLength: 26, maxLength: 26}}
    ExportId: {name: export, in: path, required: true, schema: {type: string, minLength: 26, maxLength: 26}}
    DeletionId: {name: deletion, in: path, required: true, schema: {type: string, minLength: 26, maxLength: 26}}
...
  schemas:
    Ulid: {type: string, minLength: 26, maxLength: 26, pattern: '^[0-9A-HJKMNP-TV-Z]{26}$'}
    MoneyMinor: {type: integer, minimum: 0, description: Amount in the currency minor unit; BDT uses paisa.}
    Error: {type: object, required: [code, message], properties: {code: {type: string}, field: {type: [string, 'null']}, message: {type: string}}}
    Envelope: {type: object, required: [data, meta, errors], properties: {data: {}, meta: {type: object, additionalProperties: true}, errors: {type: array, items: {$ref: '#/components/schemas/Error'}}}}
...
    PurchaseLineInput: {type: object, required: [inventory_item_id, quantity, unit_cost_minor], properties: {inventory_item_id: {$ref: '#/components/schemas/Ulid'}, quantity: {type: number, exclusiveMinimum: 0}, unit_cost_minor: {$ref: '#/components/schemas/MoneyMinor'}}}
    PurchaseInput: {type: object, required: [items], properties: {supplier_id: {oneOf: [{$ref: '#/components/schemas/Ulid'}, {type: 'null'}]}, items: {type: array, minItems: 1, items: {$ref: '#/components/schemas/PurchaseLineInput'}}}}
    SaleLineInput: {type: object, required: [inventory_item_id, quantity, unit_price_minor], properties: {inventory_item_id: {$ref: '#/components/schemas/Ulid'}, quantity: {type: number, exclusiveMinimum: 0}, unit_price_minor: {$ref: '#/components/schemas/MoneyMinor'}}}
    SaleInput: {type: object, required: [items], properties: {customer_id: {oneOf: [{$ref: '#/components/schemas/Ulid'}, {type: 'null'}]}, sale_type: {type: string, enum: [direct, fabric]}, discount_minor: {$ref: '#/components/schemas/MoneyMinor'}, paid_minor: {$ref: '#/components/schemas/MoneyMinor'}, payment_method: {type: string, enum: [cash, card, mobile_banking, bank]}, items: {type: array, minItems: 1, items: {$ref: '#/components/schemas/SaleLineInput'}}}}
    RentalLineInput: {type: object, required: [inventory_item_id, quantity], properties: {inventory_item_id: {$ref: '#/components/schemas/Ulid'}, quantity: {type: number, exclusiveMinimum: 0}, condition_out: {type: [string, 'null'], maxLength: 255}}}
    RentalInput: {type: object, required: [customer_id, starts_on, due_on, rent_minor, items], properties: {customer_id: {$ref: '#/components/schemas/Ulid'}, starts_on: {type: string, format: date}, due_on: {type: string, format: date}, rent_minor: {$ref: '#/components/schemas/MoneyMinor'}, deposit_minor: {$ref: '#/components/schemas/MoneyMinor'}, items: {type: array, minItems: 1, items: {$ref: '#/components/schemas/RentalLineInput'}}}}
    RentalReturnInput: {type: object, properties: {conditions: {type: object, additionalProperties: {type: string, maxLength: 255}}, damage_charge_minor: {$ref: '#/components/schemas/MoneyMinor'}, settlement_note: {type: [string, 'null'], maxLength: 500}}}
  requestBodies:
...
    Purchase: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/PurchaseInput'}}}}
    Sale: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/SaleInput'}}}}
    Rental: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/RentalInput'}}}}
    RentalReturn: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/RentalReturnInput'}}}}
  responses:
    Success: {description: Successful envelope, content: {application/json: {schema: {$ref: '#/components/schemas/Envelope'}}}}
...